+++ src/main/java/me/aco/marketplace_spring_ca/application/usecases/item/command/UpdateItemCommand.java
@@ -1,12 +1,14 @@
 package me.aco.marketplace_spring_ca.application.usecases.item.command;
 
+import java.math.BigDecimal;
+
 public record UpdateItemCommand(
         Long id,
         String name,
         String description,
-        double price,
-        long typeId,
-        long sellerId
+        BigDecimal price,
+        Long typeId,
+        Long sellerId
 ) {
         public static UpdateItemCommand withId(long id, UpdateItemCommand command) {
             return new UpdateItemCommand(
